Improving ENAS...
Checkpoint for the code. Working on ENAS algorithm to be more efficient, in this sense, the original code has been improved by adding MaxPool and AvgPool layer to the search space of the architectures, also the Optimizers Class has been improved to be more efficient. On the other hand, in FL Basic the data augmentation that was performed in ENAS has been implemented to have a fair comparison.\n For future changes, there is a new idea. Work on instead of partition the CIFAR-10 dataset between the clients, each one will have the same number of examples, but due to the data augmentation they will have different data. Furthermore, for FedNAS the idea is to train ENAS for a number of rounds and then fix the architecture to train efficiently with all the dataset instead of just using some batches for validation and test. In this sense, the system should perform better, lets see...