Solving problems...
Fixed problems with test accuracy, the problem was that the test images were not divided by 255 so there was an inconsistency between the range of the data values between training set and test set. On the other hand, the balance of data has been checked and the data is balanced so all is working properly.