Update FL-ENAS
Changes Performed: 1) Plot the results of training of the system, both locally and globally. 2) Change Cosine Scheduler to match with Fl-Basic, this made the performace work better, however I sholud fix when is properly to restart the Scheduler so the system would perform better. 3) Instead of using FedAvd, Custom Strategy was introduced which is the same as Fl-basic. 4) Load Data Function in Task file has been changed to be more efficient as in Fl-Basic, in this sense only training set is augmented and validation and test is not. Next step should be improve the system and smooth the plots to make it prettier for the Paper