Update FL-ENAS
Changes made: 1) Training with 4 clients. 2) Training for 20 local epoch 5 rounds. 3) solved the problem that in each round ENAS start the training from 0 and not form the previous status of the previous round, now the weights are shared the way it has to be. 4) Controller weights are saved manually due to problems with flower, this should be fixed in the future. 5) Adding the Plot of the result architecture proposed from ENAS